dot-files reloaded
This changes everything. Now there is the concept of a module, which
is a directory in the project root dir. Each module may have the
following special directories:

  1. tags: contains empty files that "tag" the current module;

  2. link: the files to create during installation;

  3. hook: defines a couple of hook (currently pre and post install);

So instead of installing everything, dot-install install only modules
that match one or more tags (OR'ed), and each module has control of
what to do.

The dot-install script has also moved into its own project now.